Move files to new structure
