+++ src/assets/components/MyModule/index.js
