+++ src/assets/components/ReactModule/index.css
@@ -1,4 +1,4 @@
-.example {
+.ReactModule {
   color: tomato;
   padding: 0.5rem;
   margin: 0.5rem;
