+++ src/index.html
@@ -4,7 +4,7 @@
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <title>Pura</title>
-  <link rel="stylesheet" href="/_compiled/styles.css">
+  <link rel="stylesheet" href="/_compiled/app.css">
 </head>
 <body>
   <h1>Pura</h1>
@@ -12,8 +12,8 @@
 
   <h2>Responsive images using Picturefill</h2>
   <picture>
-    <source srcset="/assets/images/placeholder--large.jpg" media="(min-width: 450px)">
-    <img srcset="/assets/images/placeholder--small.jpg" alt="A test image">
+    <source srcset="/images/placeholder--large.jpg" media="(min-width: 450px)">
+    <img srcset="/images/placeholder--small.jpg" alt="A test image">
   </picture>
 
   <div data-my-module></div>
