+++ tasks/subtasks/lint.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env sh
 
-./node_modules/.bin/stylelint './src/assets/styles/**/*.css' --color &&
-./node_modules/.bin/eslint './src/assets/scripts/**/*.js' --color
+./node_modules/.bin/stylelint './src/assets/**/*.css' --color &&
+./node_modules/.bin/eslint './src/assets/**/*.js' --color
