+++ tasks/subtasks/stylesProd.sh
@@ -2,9 +2,9 @@
 
 ./node_modules/.bin/postcss\
   --config ./tasks/configs/postcss.config.js\
-  --output ./src/_compiled/styles.css\
-  ./src/assets/styles/styles.css
+  --output ./src/_compiled/app.css\
+  ./src/assets/app.css
 
 ./node_modules/.bin/cleancss\
-  src/_compiled/styles.css\
-  --output src/_compiled/styles.css
+  src/_compiled/app.css\
+  --output src/_compiled/app.css
