+++ tasks/subtasks/stylesWatch.sh
@@ -2,6 +2,6 @@
 
 ./node_modules/.bin/postcss\
   --config ./tasks/configs/postcss.config.js\
-  --output ./src/_compiled/styles.css\
-  ./src/assets/styles/styles.css\
+  --output ./src/_compiled/app.css\
+  ./src/assets/app.css\
   --watch
