+++ webpack.config.js
@@ -1,32 +1,32 @@
-const webpack = require('webpack');
+const webpack = require("webpack");
 
 const PROD = process.env.NODE_ENV || 0;
 
 module.exports = {
-  devtool: PROD ? false : 'eval-cheap-module-source-map',
+  devtool: PROD ? false : "eval-cheap-module-source-map",
 
   entry: {
-    app: './src/assets/scripts/app.js',
-    vendor: ['picturefill']
+    app: "./src/assets/app.js",
+    vendor: ["picturefill"]
   },
 
   output: {
-    path: __dirname + '/src/_compiled',
-    publicPath: '/_compiled/',
-    filename: '[name].js',
-    chunkFilename: '_chunk/[name]_[chunkhash].js'
+    path: __dirname + "/src/_compiled",
+    publicPath: "/_compiled/",
+    filename: "[name].js",
+    chunkFilename: "_chunk/[name]_[chunkhash].js"
   },
 
   plugins: [
     new webpack.optimize.UglifyJsPlugin({
-      compress: PROD ? true : false,
+      compress: PROD,
       output: {
         comments: false
       }
     }),
 
     new webpack.optimize.CommonsChunkPlugin({
-      name: 'app',
+      name: "app",
       children: true,
       minChunks: 2
     })
@@ -37,7 +37,7 @@ module.exports = {
       {
         test: /\.js$/,
         exclude: /node_modules/,
-        loader: 'babel-loader'
+        loader: "babel-loader"
       }
     ]
   }
